tarama_sistemi/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
geopy==2.4.1
tenacity==8.2.3
loguru==0.7.2
PyGithub==2.1.1
fake-useragent==1.5.1
selenium==4.18.1
webdriver-manager==4.0.1